fix Value::home integer overflow
A method with a large number of local variable slots may imply offsets
larger than a int8_t can represent, so we now use an int16_t instead.

+++ src/compiler.cpp
@@ -326,7 +326,7 @@ class Value: public Compiler::Operand {
   Site* target;
   Value* buddy;
   Value* nextWord;
-  int8_t home;
+  int16_t home;
   ValueType type;
   uint8_t wordIndex;
 };
